+++ python/sglang/srt/layers/moe/topk.py
@@ -699,7 +699,7 @@ def grouped_topk_kunpeng(
     )
 
     topk_weights = token_weights
-    topk_ids = token_ids.to(torch.int32)
+    topk_ids = token_ids
 
     if num_fused_shared_experts:
         topk_ids[:, -1] = torch.randint(
@@ -1168,7 +1168,7 @@ def biased_grouped_topk_kunpeng(
     )
 
     topk_weights = token_weights
-    topk_ids = token_ids.to(torch.int32)
+    topk_ids = token_ids
 
     if num_fused_shared_experts:
         topk_ids[:, -1] = torch.randint(
